+++ packages/models-library/src/models_library/api_schemas_directorv2/errors.py
@@ -1,15 +1,13 @@
-from typing import List, Optional
-
 from pydantic import BaseModel, Field
 
 
 class Error(BaseModel):
-    code: Optional[str] = Field(None, description="Server Exception")
+    code: str | None = Field(None, description="Server Exception")
 
 
 class ErrorType(BaseModel):
     message: str = Field(..., description="Error message")
-    errors: Optional[List[Error]] = None
+    errors: list[Error] | None = None
     status: int = Field(..., description="Error code")
