+++ services/director-v2/src/simcore_service_director_v2/api/routes/computations_tasks.py
@@ -11,6 +11,7 @@ from typing import NamedTuple
 
 import networkx as nx
 from fastapi import APIRouter, Depends, HTTPException
+from models_library.api_schemas_directorv2.comp_tasks import TaskLogFileGet
 from models_library.projects import ProjectID
 from models_library.projects_nodes_io import NodeID
 from models_library.users import UserID
@@ -21,7 +22,6 @@ from starlette import status
 
 from ...models.domains.comp_pipelines import CompPipelineAtDB
 from ...models.domains.comp_tasks import CompTaskAtDB
-from ...models.schemas.comp_tasks import TaskLogFileGet
 from ...modules.db.repositories.comp_pipelines import CompPipelinesRepository
 from ...modules.db.repositories.comp_tasks import CompTasksRepository
 from ...utils.dask import get_service_log_file_download_link
