+++ services/director-v2/src/simcore_service_director_v2/modules/director_v0.py
@@ -13,6 +13,7 @@ from fastapi import FastAPI, HTTPException, status
 from models_library.api_schemas_directorv2.dynamic_services import (
     RunningDynamicServiceDetails,
 )
+from models_library.api_schemas_directorv2.services import ServiceExtras
 from models_library.projects import ProjectID
 from models_library.projects_nodes import NodeID
 from models_library.service_settings_labels import SimcoreServiceLabels
@@ -21,7 +22,6 @@ from models_library.users import UserID
 from servicelib.logging_utils import log_decorator
 
 from ..core.settings import DirectorV0Settings
-from ..models.schemas.services import ServiceExtras
 from ..utils.client_decorators import handle_errors, handle_retry
 from ..utils.clients import unenvelope_or_raise_error
