+++ services/director-v2/src/simcore_service_director_v2/utils/dask.py
@@ -20,6 +20,7 @@ from dask_task_models_library.container_tasks.protocol import (
     ContainerLabelsDict,
 )
 from fastapi import FastAPI
+from models_library.api_schemas_directorv2.services import NodeRequirements
 from models_library.clusters import ClusterID
 from models_library.docker import StandardSimcoreDockerLabels
 from models_library.errors import ErrorDict
@@ -48,7 +49,6 @@ from ..core.errors import (
 )
 from ..models.domains.comp_runs import MetadataDict
 from ..models.domains.comp_tasks import Image
-from ..models.schemas.services import NodeRequirements
 from ..modules.osparc_variables_substitutions import (
     resolve_and_substitute_session_variables_in_specs,
     substitute_vendor_secrets_in_specs,
