+++ services/director-v2/tests/integration/02/test_dynamic_sidecar_nodeports_integration.py
@@ -37,6 +37,7 @@ from helpers.shared_comp_utils import (
     assert_and_wait_for_pipeline_status,
     assert_computation_task_out_obj,
 )
+from models_library.api_schemas_directorv2.comp_tasks import ComputationGet
 from models_library.api_schemas_directorv2.constants import (
     DYNAMIC_SIDECAR_SERVICE_PREFIX,
 )
@@ -79,7 +80,6 @@ from simcore_sdk import node_ports_v2
 from simcore_sdk.node_data import data_manager
 from simcore_sdk.node_ports_v2 import DBManager, Nodeports, Port
 from simcore_service_director_v2.core.settings import AppSettings, RCloneSettings
-from simcore_service_director_v2.models.schemas.comp_tasks import ComputationGet
 from sqlalchemy.dialects.postgresql import insert as pg_insert
 from tenacity._asyncio import AsyncRetrying
 from tenacity.retry import retry_if_exception_type
