+++ services/director-v2/tests/unit/with_dbs/test_api_route_computations.py
@@ -18,6 +18,11 @@ import pytest
 import respx
 from faker import Faker
 from fastapi import FastAPI
+from models_library.api_schemas_directorv2.comp_tasks import (
+    ComputationCreate,
+    ComputationGet,
+)
+from models_library.api_schemas_directorv2.services import ServiceExtras
 from models_library.basic_types import VersionStr
 from models_library.clusters import DEFAULT_CLUSTER_ID, Cluster, ClusterID
 from models_library.projects import ProjectAtDB
@@ -40,11 +45,6 @@ from simcore_postgres_database.models.comp_tasks import NodeClass
 from simcore_service_director_v2.models.domains.comp_pipelines import CompPipelineAtDB
 from simcore_service_director_v2.models.domains.comp_runs import CompRunsAtDB
 from simcore_service_director_v2.models.domains.comp_tasks import CompTaskAtDB
-from simcore_service_director_v2.models.schemas.comp_tasks import (
-    ComputationCreate,
-    ComputationGet,
-)
-from simcore_service_director_v2.models.schemas.services import ServiceExtras
 from starlette import status
 
 pytest_simcore_core_services_selection = ["postgres", "rabbit"]
