+++ services/director-v2/tests/unit/with_dbs/test_api_route_computations_tasks.py
@@ -10,6 +10,7 @@ import httpx
 import pytest
 from faker import Faker
 from fastapi import FastAPI, status
+from models_library.api_schemas_directorv2.comp_tasks import TaskLogFileGet
 from models_library.projects import ProjectAtDB, ProjectID
 from models_library.projects_nodes_io import NodeID
 from models_library.users import UserID
@@ -18,7 +19,6 @@ from pytest_simcore.helpers.typing_env import EnvVarsDict
 from simcore_service_director_v2.core.settings import AppSettings
 from simcore_service_director_v2.models.domains.comp_pipelines import CompPipelineAtDB
 from simcore_service_director_v2.models.domains.comp_tasks import CompTaskAtDB
-from simcore_service_director_v2.models.schemas.comp_tasks import TaskLogFileGet
 
 pytest_simcore_core_services_selection = [
     "postgres",
